Changed the organization of the repository
Added the bin, src, and inc directories. Also moved everything to
utilities directory, so it will nicely merge to main later on.

The KITT class is now moved to the package, and all of the files that
require this package append the inc directory to their path and import
the class. Everything has to be tested.

Bin directory is for later use.